% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{BackupConfiguration}
\alias{BackupConfiguration}
\title{Cloud SQL Administration API Objects 
Creates and configures Cloud SQL instances, which provide fully-managed MySQL databases.}
\usage{
BackupConfiguration(binaryLogEnabled = NULL, enabled = NULL, id = NULL,
  startTime = NULL)
}
\arguments{
\item{binaryLogEnabled}{Whether binary log is enabled}

\item{enabled}{Whether this configuration is enabled}

\item{id}{Identifier for this configuration}

\item{startTime}{Start time for the daily backup configuration in UTC timezone in the 24 hour format - HH:MM}
}
\value{
BackupConfiguration object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:47:51
filename: /Users/mark/dev/R/autoGoogleAPI/googlesqladminv1beta3.auto/R/sqladmin_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
BackupConfiguration Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance backup configuration.
}

