% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{BackupRun}
\alias{BackupRun}
\title{BackupRun Object}
\usage{
BackupRun(backupConfiguration = NULL, dueTime = NULL, endTime = NULL,
  enqueuedTime = NULL, error = NULL, instance = NULL, startTime = NULL,
  status = NULL)
}
\arguments{
\item{backupConfiguration}{Backup Configuration identifier}

\item{dueTime}{The due time of this run in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{endTime}{The time the backup operation completed in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{enqueuedTime}{The time the run was enqueued in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{error}{Information about why the backup operation failed}

\item{instance}{Name of the database instance}

\item{startTime}{The time the backup operation actually started in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{status}{The status of this run}
}
\value{
BackupRun object
}
\description{
BackupRun Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A database instance backup run resource.
}
