% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{BackupRunsListResponse}
\alias{BackupRunsListResponse}
\title{BackupRunsListResponse Object}
\usage{
BackupRunsListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{A list of backup runs in reverse chronological order of the enqueued time}

\item{nextPageToken}{The continuation token, used to page through large result sets}
}
\value{
BackupRunsListResponse object
}
\description{
BackupRunsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Backup run list results.
}

