% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{CloneContext}
\alias{CloneContext}
\title{CloneContext Object}
\usage{
CloneContext(binLogCoordinates = NULL, destinationInstanceName = NULL,
  sourceInstanceName = NULL)
}
\arguments{
\item{binLogCoordinates}{Binary log coordinates, if specified, indentify the position up to which the source instance should be cloned}

\item{destinationInstanceName}{Name of the Cloud SQL instance to be created as a clone}

\item{sourceInstanceName}{Name of the Cloud SQL instance to be cloned}
}
\value{
CloneContext object
}
\description{
CloneContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance clone context.
}

