% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{DatabaseFlags}
\alias{DatabaseFlags}
\title{DatabaseFlags Object}
\usage{
DatabaseFlags(name = NULL, value = NULL)
}
\arguments{
\item{name}{The name of the flag}

\item{value}{The value of the flag}
}
\value{
DatabaseFlags object
}
\description{
DatabaseFlags Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
MySQL flags for Cloud SQL instances.
}
