% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{ExportContext}
\alias{ExportContext}
\title{ExportContext Object}
\usage{
ExportContext(database = NULL, table = NULL, uri = NULL)
}
\arguments{
\item{database}{Databases (for example, guestbook) from which the export is made}

\item{table}{Tables to export, or that were exported, from the specified database}

\item{uri}{The path to the file in Google Cloud Storage where the export will be stored, or where it was already stored}
}
\value{
ExportContext object
}
\description{
ExportContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance export context.
}
