% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{ImportContext}
\alias{ImportContext}
\title{ImportContext Object}
\usage{
ImportContext(database = NULL, uri = NULL)
}
\arguments{
\item{database}{The database (for example, guestbook) to which the import is made}

\item{uri}{A path to the MySQL dump file in Google Cloud Storage from which the import is made}
}
\value{
ImportContext object
}
\description{
ImportContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance import context.
}
