% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstanceOperation}
\alias{InstanceOperation}
\title{InstanceOperation Object}
\usage{
InstanceOperation(endTime = NULL, enqueuedTime = NULL, error = NULL,
  exportContext = NULL, importContext = NULL, instance = NULL,
  operation = NULL, operationType = NULL, startTime = NULL,
  state = NULL, userEmailAddress = NULL)
}
\arguments{
\item{endTime}{The time this operation finished in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{enqueuedTime}{The time this operation was enqueued in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{error}{The error(s) encountered by this operation}

\item{exportContext}{The context for export operation, if applicable}

\item{importContext}{The context for import operation, if applicable}

\item{instance}{Name of the database instance}

\item{operation}{An identifier that uniquely identifies the operation}

\item{operationType}{The type of the operation}

\item{startTime}{The time this operation actually started in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{state}{The state of an operation}

\item{userEmailAddress}{The email address of the user who initiated this operation}
}
\value{
InstanceOperation object
}
\description{
InstanceOperation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Operations resource contains information about database instance operations such as create, delete, and restart. Operations resources are created in response to operations that were initiated; you never create them directly.
}
