% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesCloneRequest}
\alias{InstancesCloneRequest}
\title{InstancesCloneRequest Object}
\usage{
InstancesCloneRequest(cloneContext = NULL)
}
\arguments{
\item{cloneContext}{Contains details about the clone operation}
}
\value{
InstancesCloneRequest object
}
\description{
InstancesCloneRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance clone request.
}
\seealso{
Other InstancesCloneRequest functions: \code{\link{sql.instances.clone}}
}

