% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesDeleteResponse}
\alias{InstancesDeleteResponse}
\title{InstancesDeleteResponse Object}
\usage{
InstancesDeleteResponse(operation = NULL)
}
\arguments{
\item{operation}{An identifier that uniquely identifies the operation}
}
\value{
InstancesDeleteResponse object
}
\description{
InstancesDeleteResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance delete response.
}

