% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstanceSetRootPasswordRequest}
\alias{InstanceSetRootPasswordRequest}
\title{InstanceSetRootPasswordRequest Object}
\usage{
InstanceSetRootPasswordRequest(setRootPasswordContext = NULL)
}
\arguments{
\item{setRootPasswordContext}{Set Root Password Context}
}
\value{
InstanceSetRootPasswordRequest object
}
\description{
InstanceSetRootPasswordRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance set root password request.
}
\seealso{
Other InstanceSetRootPasswordRequest functions: \code{\link{sql.instances.setRootPassword}}
}
