% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesExportResponse}
\alias{InstancesExportResponse}
\title{InstancesExportResponse Object}
\usage{
InstancesExportResponse(operation = NULL)
}
\arguments{
\item{operation}{An identifier that uniquely identifies the operation}
}
\value{
InstancesExportResponse object
}
\description{
InstancesExportResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance export response.
}

