% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesImportRequest}
\alias{InstancesImportRequest}
\title{InstancesImportRequest Object}
\usage{
InstancesImportRequest(importContext = NULL)
}
\arguments{
\item{importContext}{Contains details about the import operation}
}
\value{
InstancesImportRequest object
}
\description{
InstancesImportRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance import request.
}
\seealso{
Other InstancesImportRequest functions: \code{\link{sql.instances.import}}
}

