% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesImportResponse}
\alias{InstancesImportResponse}
\title{InstancesImportResponse Object}
\usage{
InstancesImportResponse(operation = NULL)
}
\arguments{
\item{operation}{An identifier that uniquely identifies the operation}
}
\value{
InstancesImportResponse object
}
\description{
InstancesImportResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance import response.
}

