% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesInsertResponse}
\alias{InstancesInsertResponse}
\title{InstancesInsertResponse Object}
\usage{
InstancesInsertResponse(operation = NULL)
}
\arguments{
\item{operation}{An identifier that uniquely identifies the operation}
}
\value{
InstancesInsertResponse object
}
\description{
InstancesInsertResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance insert response.
}

