% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesPromoteReplicaResponse}
\alias{InstancesPromoteReplicaResponse}
\title{InstancesPromoteReplicaResponse Object}
\usage{
InstancesPromoteReplicaResponse(operation = NULL)
}
\arguments{
\item{operation}{An identifier that uniquely identifies the operation}
}
\value{
InstancesPromoteReplicaResponse object
}
\description{
InstancesPromoteReplicaResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database promote read replica response.
}

