% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesRestartResponse}
\alias{InstancesRestartResponse}
\title{InstancesRestartResponse Object}
\usage{
InstancesRestartResponse(operation = NULL)
}
\arguments{
\item{operation}{An identifier that uniquely identifies the operation}
}
\value{
InstancesRestartResponse object
}
\description{
InstancesRestartResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance restart response.
}

