% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesUpdateResponse}
\alias{InstancesUpdateResponse}
\title{InstancesUpdateResponse Object}
\usage{
InstancesUpdateResponse(operation = NULL)
}
\arguments{
\item{operation}{An identifier that uniquely identifies the operation}
}
\value{
InstancesUpdateResponse object
}
\description{
InstancesUpdateResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance update response.
}
