% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{IpConfiguration}
\alias{IpConfiguration}
\title{IpConfiguration Object}
\usage{
IpConfiguration(authorizedNetworks = NULL, enabled = NULL,
  requireSsl = NULL)
}
\arguments{
\item{authorizedNetworks}{The list of external networks that are allowed to connect to the instance using the IP}

\item{enabled}{Whether the instance should be assigned an IP address or not}

\item{requireSsl}{Whether the mysqld should default to 'REQUIRE X509' for users connecting over IP}
}
\value{
IpConfiguration object
}
\description{
IpConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
IP Management configuration.
}

