% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{IpMapping}
\alias{IpMapping}
\title{IpMapping Object}
\usage{
IpMapping(ipAddress = NULL, timeToRetire = NULL)
}
\arguments{
\item{ipAddress}{The IP address assigned}

\item{timeToRetire}{The due time for this IP to be retired in RFC 3339 format, for example 2012-11-15T16:19:00}
}
\value{
IpMapping object
}
\description{
IpMapping Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance IP Mapping.
}

