% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{LocationPreference}
\alias{LocationPreference}
\title{LocationPreference Object}
\usage{
LocationPreference(followGaeApplication = NULL, zone = NULL)
}
\arguments{
\item{followGaeApplication}{The App Engine application to follow, it must be in the same region as the Cloud SQL instance}

\item{zone}{The preferred Compute Engine zone (e}
}
\value{
LocationPreference object
}
\description{
LocationPreference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Preferred location. This specifies where a Cloud SQL instance should preferably be located, either in a specific Compute Engine zone, or co-located with an App Engine application. Note that if the preferred location is not available, the instance will be located as close as possible within the region. Only one location may be specified.
}

