% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{SetRootPasswordContext}
\alias{SetRootPasswordContext}
\title{SetRootPasswordContext Object}
\usage{
SetRootPasswordContext(password = NULL)
}
\arguments{
\item{password}{The password for the root user}
}
\value{
SetRootPasswordContext object
}
\description{
SetRootPasswordContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance set root password context.
}
