% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{Settings}
\alias{Settings}
\title{Settings Object}
\usage{
Settings(activationPolicy = NULL, authorizedGaeApplications = NULL,
  backupConfiguration = NULL, databaseFlags = NULL,
  databaseReplicationEnabled = NULL, ipConfiguration = NULL,
  locationPreference = NULL, pricingPlan = NULL, replicationType = NULL,
  settingsVersion = NULL, tier = NULL)
}
\arguments{
\item{activationPolicy}{The activation policy for this instance}

\item{authorizedGaeApplications}{The App Engine app IDs that can access this instance}

\item{backupConfiguration}{The daily backup configuration for the instance}

\item{databaseFlags}{The database flags passed to the instance at startup}

\item{databaseReplicationEnabled}{Configuration specific to read replica instance}

\item{ipConfiguration}{The settings for IP Management}

\item{locationPreference}{The location preference settings}

\item{pricingPlan}{The pricing plan for this instance}

\item{replicationType}{The type of replication this instance uses}

\item{settingsVersion}{The version of instance settings}

\item{tier}{The tier of service for this instance, for example D1, D2}
}
\value{
Settings object
}
\description{
Settings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance settings.
}
