% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_functions.R
\name{sql.instances.clone}
\alias{sql.instances.clone}
\title{Creates a Cloud SQL instance as a clone of a source instance.}
\usage{
sql.instances.clone(InstancesCloneRequest, project)
}
\arguments{
\item{InstancesCloneRequest}{The \link{InstancesCloneRequest} object to pass to this method}

\item{project}{Project ID of the source as well as the clone Cloud SQL instance}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/sqlservice.admin
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/sqlservice.admin)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/sql/docs/reference/latest}{Google Documentation}

Other InstancesCloneRequest functions: \code{\link{InstancesCloneRequest}}
}

