% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_functions.R
\name{sql.instances.patch}
\alias{sql.instances.patch}
\title{Updates the settings of a Cloud SQL instance. This method supports patch semantics.}
\usage{
sql.instances.patch(DatabaseInstance, project, instance)
}
\arguments{
\item{DatabaseInstance}{The \link{DatabaseInstance} object to pass to this method}

\item{project}{Project ID of the project that contains the instance}

\item{instance}{Cloud SQL instance ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/sqlservice.admin
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/sqlservice.admin)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/sql/docs/reference/latest}{Google Documentation}

Other DatabaseInstance functions: \code{\link{DatabaseInstance}},
  \code{\link{sql.instances.insert}},
  \code{\link{sql.instances.update}}
}

