% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_functions.R
\docType{package}
\name{sqladmin_googleAuthR}
\alias{sqladmin_googleAuthR}
\alias{sqladmin_googleAuthR-package}
\title{Cloud SQL Administration API
Creates and configures Cloud SQL instances, which provide fully-managed MySQL databases.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:16:09
filename: /Users/mark/dev/R/autoGoogleAPI/googlesqladminv1beta3.auto/R/sqladmin_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/sqlservice.admin
}
}
