% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{SslCert}
\alias{SslCert}
\title{SslCert Object}
\usage{
SslCert(cert = NULL, certSerialNumber = NULL, commonName = NULL,
  createTime = NULL, expirationTime = NULL, instance = NULL,
  sha1Fingerprint = NULL)
}
\arguments{
\item{cert}{PEM representation}

\item{certSerialNumber}{Serial number, as extracted from the certificate}

\item{commonName}{User supplied name}

\item{createTime}{Time when the certificate was created}

\item{expirationTime}{Time when the certificate expires}

\item{instance}{Name of the database instance}

\item{sha1Fingerprint}{Sha1 Fingerprint}
}
\value{
SslCert object
}
\description{
SslCert Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SslCerts Resource
}
