% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{SslCertDetail}
\alias{SslCertDetail}
\title{SslCertDetail Object}
\usage{
SslCertDetail(certInfo = NULL, certPrivateKey = NULL)
}
\arguments{
\item{certInfo}{The public information about the cert}

\item{certPrivateKey}{The private key for the client cert, in pem format}
}
\value{
SslCertDetail object
}
\description{
SslCertDetail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SslCertDetail.
}

