% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{SslCertsDeleteResponse}
\alias{SslCertsDeleteResponse}
\title{SslCertsDeleteResponse Object}
\usage{
SslCertsDeleteResponse(operation = NULL)
}
\arguments{
\item{operation}{An identifier that uniquely identifies the operation}
}
\value{
SslCertsDeleteResponse object
}
\description{
SslCertsDeleteResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SslCert delete response.
}

