% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{SslCertsInsertRequest}
\alias{SslCertsInsertRequest}
\title{SslCertsInsertRequest Object}
\usage{
SslCertsInsertRequest(commonName = NULL)
}
\arguments{
\item{commonName}{User supplied name}
}
\value{
SslCertsInsertRequest object
}
\description{
SslCertsInsertRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SslCerts insert request.
}
\seealso{
Other SslCertsInsertRequest functions: \code{\link{sql.sslCerts.insert}}
}

