% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{SslCertsListResponse}
\alias{SslCertsListResponse}
\title{SslCertsListResponse Object}
\usage{
SslCertsListResponse(items = NULL)
}
\arguments{
\item{items}{List of client certificates for the instance}
}
\value{
SslCertsListResponse object
}
\description{
SslCertsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SslCerts list response.
}

