% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{Tier}
\alias{Tier}
\title{Tier Object}
\usage{
Tier(DiskQuota = NULL, RAM = NULL, region = NULL, tier = NULL)
}
\arguments{
\item{DiskQuota}{The maximum disk size of this tier in bytes}

\item{RAM}{The maximum RAM usage of this tier in bytes}

\item{region}{The applicable regions for this tier}

\item{tier}{An identifier for the service tier, for example D1, D2 etc}
}
\value{
Tier object
}
\description{
Tier Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Google Cloud SQL service tier resource.
}
