% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{TiersListResponse}
\alias{TiersListResponse}
\title{TiersListResponse Object}
\usage{
TiersListResponse(items = NULL)
}
\arguments{
\item{items}{List of tiers}
}
\value{
TiersListResponse object
}
\description{
TiersListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Tiers list response.
}

